ImagePreprocessing/requirements.txt
torch>=1.5
numpy>=1.17
pillow>=5.3.0
matplotlib>=1.5.1
scikit-image>=0.13.0
scikit-learn>=0.20.1
nibabel>=2.3.0